Unifies three lines of work onto one integration branch:
  * the 133 upstream Asterinas commits the ports were built on,
  * the AArch64 (ARM64) and 32-bit ARMv7-A ports (boot to userspace; ARM runs
    stock static Linux ELF programs), and
  * the 54-commit `asterkube` Kubernetes-node stack (netfilter/kube-proxy
    Service NAT, seccomp BPF, rootless user namespaces, astromac tenant MAC,
    cgroup/procfs/sysfs knobs the kubelet needs, netlink, containerd/ctr, the
    zstd node image).

Resolved 20 merge conflicts by *combining* both sides, not choosing one:
  * security/lsm: keep upstream's `capability` module + `LsmCapabilityHook`
    AND asterkube's `astromac` module + file/signal/socket hooks; every module
    now implements all five hook traits.
  * net: keep both `ReceiveBehavior` (upstream) and `IcmpSocket`/nat exports
    (asterkube); netlink attr ordering and `SocketType(i32)` from asterkube.
  * procfs: union /proc/sys/{fs,kernel,net,vm} and the vm knobs
    (mmap_min_addr + overcommit_memory/panic_on_oom).
  * process: constructor keeps upstream `user_ns` and asterkube `pid_ns`/`vpid`;
    execve combines ambient-capset clearing with `nosuid` enforcement.
  * fs: rootfs streams gzip (zune-inflate) AND zstd through one BoxedReader;
    pivot_root uses the topology-guard API with asterkube's namespace-root case.

Then fixed 18 semantic breaks where asterkube code (written against the older
base) met the newer upstream APIs: `iface.name()` now returns `&CStr`,
`IpIface::new` takes `CString`, `is_seqpacket()` is a method, `generate_control`
takes a `ReceiveBehavior`, the netlink/netfilter socket ctor takes a `SockType`,
and `crate::arch::init_late` (the x86 ACPI power-button arm) gets no-op stubs on
AArch64/ARM.

Validated: compiles on x86_64 (K8s target), aarch64, and arm; the ARM kernel
still boots to userspace and runs a real static musl ELF to exit(0).

Kubernetes-on-AArch64 (running the node image on the ARM64 kernel) remains a
follow-up: this unifies the code but the node datapath is exercised on x86 only.

Fully AI-assisted; lands on an integration branch, main untouched.

+++ kernel/src/net/socket/util/options.rs
@@ -116,7 +116,7 @@ impl SocketOptionSet {
                 socket_reuse_addr.set(reuse_addr);
             }
             socket_type @ SocketType => {
-                socket_type.set(socket.socket_type());
+                socket_type.set(socket.socket_type() as i32);
             }
             socket_broadcast @ Broadcast => {
                 let broadcast = self.broadcast();
